+++ plugins/auth-backend-module-microsoft-provider/src/index.ts
@@ -21,5 +21,6 @@
  */
 
 export { microsoftAuthenticator } from './authenticator';
-export { authModuleMicrosoftProvider } from './module';
+export { authModuleMicrosoftProvider as default } from './module';
 export { microsoftSignInResolvers } from './resolvers';
+export * from './deprecated';