+++ plugins/azure-devops/src/components/PullRequestTable/PullRequestTable.tsx
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import { Box, Chip } from '@material-ui/core';
+import Box from '@material-ui/core/Box';
+import Chip from '@material-ui/core/Chip';
 import {
   Link,
   ResponseErrorPanel,