+++ packages/backend-openapi-utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/backend-openapi-utils
 
+## 0.1.2
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/errors@1.2.3
+
+## 0.1.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9-next.2
+
 ## 0.1.2-next.1
 
 ### Patch Changes