+++ plugins/airbrake-backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-airbrake-backend
 
+## 0.3.6
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/config@1.1.1
+
+## 0.3.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9-next.2
+  - @backstage/backend-common@0.20.1-next.2
+
 ## 0.3.6-next.1
 
 ### Patch Changes