+++ plugins/airbrake/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @backstage/plugin-airbrake
 
+## 0.3.29
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/core-components@0.13.10
+  - @backstage/core-plugin-api@1.8.2
+  - @backstage/plugin-catalog-react@1.9.3
+  - @backstage/dev-utils@1.0.26
+  - @backstage/catalog-model@1.4.3
+  - @backstage/test-utils@1.4.7
+
+## 0.3.29-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.9.3-next.2
+  - @backstage/dev-utils@1.0.26-next.2
+
 ## 0.3.29-next.1
 
 ### Patch Changes