+++ plugins/app-backend/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @backstage/plugin-app-backend
 
+## 0.3.57
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/config-loader@1.6.1
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/config@1.1.1
+  - @backstage/types@1.1.1
+  - @backstage/plugin-app-node@0.1.9
+
+## 0.3.57-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9-next.2
+  - @backstage/backend-common@0.20.1-next.2
+  - @backstage/plugin-app-node@0.1.9-next.2
+  - @backstage/config-loader@1.6.1-next.0
+
 ## 0.3.57-next.1
 
 ### Patch Changes