+++ plugins/auth-backend-module-gitlab-provider/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-auth-backend-module-gitlab-provider
 
+## 0.1.6
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/plugin-auth-node@0.4.3
+
+## 0.1.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9-next.2
+  - @backstage/plugin-auth-node@0.4.3-next.2
+
 ## 0.1.6-next.1
 
 ### Patch Changes