+++ plugins/azure-sites-backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-azure-sites-backend
 
+## 0.1.19
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/config@1.1.1
+  - @backstage/plugin-azure-sites-common@0.1.1
+
+## 0.1.19-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.20.1-next.2
+
 ## 0.1.19-next.1
 
 ### Patch Changes