+++ plugins/kubernetes-node/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-kubernetes-node
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/plugin-kubernetes-common@0.7.3
+  - @backstage/catalog-model@1.4.3
+  - @backstage/types@1.1.1
+
+## 0.1.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9-next.2
+
 ## 0.1.3-next.1
 
 ### Patch Changes