+++ plugins/opencost/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-opencost
 
+## 0.2.4
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/core-components@0.13.10
+  - @backstage/core-plugin-api@1.8.2
+
 ## 0.2.4-next.1
 
 ### Patch Changes